make ApiParser an abstract class and create EtherscanApiParser
